Rename database.py -> data_models.py
The terminology in ORMs like Peewee is more about "models" than
databases, so to fit with expectations, it seems better to rename the
file database.py
